Cleanup.

+++ backend/distributed/sfc/mars_sfc_utils.hpp
@@ -8,7 +8,6 @@
  * @author Daniel Ganellari
  */
 
-#include <_types/_uint64_t.h>
 #include "mars_base.hpp"
 #include "mars_config.hpp"
 #include "mars_globals.hpp"
@@ -24,10 +23,10 @@ template <>
 struct maxTreeLevel<unsigned> : integral_constant<unsigned, 10> {};
 
 template <>
-struct maxTreeLevel<uint64_t> : integral_constant<unsigned, 21> {};
+struct maxTreeLevel<unsigned long> : integral_constant<unsigned, 21> {};
 
 template <>
-struct maxTreeLevel<unsigned long> : integral_constant<unsigned, 21> {};
+struct maxTreeLevel<unsigned long long> : integral_constant<unsigned, 21> {};
 
 //! @brief maximum integer coordinate
 template <class KeyType>
